Conflicts:
	obdalib-core/src/main/java/it/unibz/krdb/obda/io/ModelIOManager.java
	reformulation-core/src/main/java/it/unibz/krdb/obda/owlrefplatform/core/basicoperations/UnifierUtilities.java

+++ quest-owlapi3/src/main/java/it/unibz/krdb/obda/owlrefplatform/owlapi3/QuestOWLStatement.java
@@ -356,7 +356,7 @@ public class QuestOWLStatement implements AutoCloseable {
 		}
 	}
 
-	public int getTupleCount(String query) throws OWLException {
+	public long getTupleCount(String query) throws OWLException {
 		try {
 			return st.getTupleCount(query);
 		} catch (Exception e) {
